requirements.txt
streamlit
pandas
numpy
matplotlib
plotly
requests
nltk
uvicorn
praw
fastapi
dotenv
pytest-playwright